+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.1.0
+
+* Add error handling for sync and async signals.
+* Add services as 4 argument of every action.
+
 ## v1.0.0
 
 * Initial Release.